requirements.txt
asgiref==3.6.0
boto3==1.26.93
botocore==1.29.93
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
django-storages==1.13.2
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
